requirements-dev.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-configurations==2.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-widget-tweaks==1.4.8
Pillow==7.2.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
